CLI の `.mxl` / `.mscz` file I/O 対応と ZIP 処理の共通化
## 概要

CLI で MusicXML / MuseScore の圧縮形式を file I/O に限って扱えるようにします。あわせて、既存の ZIP 読み書き処理を共通 helper に整理し、CLI と既存処理の再利用性を高めます。

## 変更内容

- CLI で `--from musicxml` の入力として `.mxl` を扱えるように変更
- CLI で `--from musescore` の入力として `.mscz` を扱えるように変更
- CLI で `--to musicxml` の出力先が `.mxl` の場合に圧縮 MusicXML を出力するように変更
- CLI で `--to musescore` の出力先が `.mscz` の場合に圧縮 MuseScore を出力するように変更
- `stdin` / `stdout` は従来どおり text-only のまま維持
- ZIP 読み書き処理を `src/ts/zip-io.ts` に共通化
- `src/ts/mxl-io.ts` と `src/ts/download-flow.ts` から ZIP 共通 helper を利用する構成に整理
- `src/ts/cli-api.ts` に CLI 向けの ZIP 入出力 helper を追加
- `scripts/mikuscore-cli.mjs` の help 文面を ZIP file I/O 対応に合わせて更新
- `README.md`、`docs/DEVELOPMENT.md`、`docs/spec/CLI_STEP1.md`、`docs/future/CLI_ROADMAP.md`、`TODO.md` を更新

## テスト

- `tests/unit/cli-api.spec.ts` に `.mxl` / `.mscz` の encode / decode テストを追加
- `tests/unit/mikuscore-cli.spec.ts` に `.mxl` / `.mscz` の CLI file input / output テストを追加

## 補足

- `index.html`、`mikuscore.html`、`src/js/main.js` は関連変更に伴う更新を含みます
- `TODO.md` には、今回の対応完了に加えて、重い `musescore-io` roundtrip テストの見直しタスクが追記されています
