+++ index.html
@@ -125,7 +125,7 @@
 <body>
   <main class="card">
     <h1>mikuscore</h1>
-    <p class="sub">Updated: 2026-04-15</p>
+    <p class="sub">Updated: 2026-04-16</p>
     <section class="lang-block" aria-label="English">
       <h2>English</h2>
       <p>
